refactor:LottoService+LottoUtil->역할에 따라 분리
